+++ multibody/multibody_tree/implicit_stribeck/BUILD.bazel
@@ -13,13 +13,6 @@ package(
     default_visibility = ["//visibility:public"],
 )
 
-drake_cc_package_library(
-    name = "contact_solvers",
-    deps = [
-        ":implicit_stribeck_solver",
-    ],
-)
-
 drake_cc_library(
     name = "implicit_stribeck_solver",
     srcs = ["implicit_stribeck_solver.cc"],
@@ -35,7 +28,6 @@ drake_cc_googletest(
     name = "implicit_stribeck_solver_test",
     deps = [
         ":implicit_stribeck_solver",
-        #        "//common/test_utilities:expect_throws_message",
     ],
 )
