+++ multibody/multibody_tree/implicit_stribeck/implicit_stribeck_solver.h
